pyproject.toml
[project]
name = "ecombot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
    "google-generativeai>=0.8.3",
]
